Fix interval type handling in REST API
Previously, interval day to second types greater than INT_MAX milliseconds
(~24 days) and interval year to month types greater than INT_MAX months
returned the wrong value due to integer overflow.

+++ presto-spi/src/main/java/com/facebook/presto/spi/type/IntervalDayTimeType.java
@@ -48,7 +48,7 @@ public final class IntervalDayTimeType
         if (block.isNull(position)) {
             return null;
         }
-        return new SqlIntervalDayTime((int) block.getLong(position, 0));
+        return new SqlIntervalDayTime(block.getLong(position, 0));
     }
 
     @Override
